# fixed sf::Image::create