@eXpl0it3r 2016-11-11
SFMLBuildMaster: Build this please.