@binary1248 2016-11-12
My proposal would look something like this:
```cpp
////////////////////////////////////////////////////////////
void Image::create(unsigned int width, unsigned int height, const Color& color)
{
    if (width && height)
    {
        // Create a new pixel buffer first for exception safety's sake
        std::vector<Uint8> newPixels(width * height * 4);

        // Fill it with the specified color
        Uint8* ptr = &newPixels[0];
        Uint8* end = ptr + newPixels.size();
        while (ptr < end)
        {
            *ptr++ = color.r;
            *ptr++ = color.g;
            *ptr++ = color.b;
            *ptr++ = color.a;
        }

        // Commit the new pixel buffer
        m_pixels.swap(newPixels);

        // Assign the new size
        m_size.x = width;
        m_size.y = height;
    }
    else
    {
        // Truncate the pixel buffer
        m_pixels.swap(std::vector<Uint8>());

        // Assign the new size
        m_size.x = 0;
        m_size.y = 0;
    }
}


////////////////////////////////////////////////////////////
void Image::create(unsigned int width, unsigned int height, const Uint8* pixels)
{
    if (pixels && width && height)
    {
        // Swap out the contents of the pixel buffer
        m_pixels.swap(std::vector<Uint8>(pixels, pixels + width * height * 4));

        // Assign the new size
        m_size.x = width;
        m_size.y = height;
    }
    else
    {
        // Truncate the pixel buffer
        m_pixels.swap(std::vector<Uint8>());

        // Assign the new size
        m_size.x = 0;
        m_size.y = 0;
    }
}
```