@Fytch 2016-11-12
@binary1248 Your proposal won't compile because swap requires an lvalue while you're trying to call it with a temporary. Change that the other way round `temp.swap(m_pixels);` and it'll work. I agree with your changes regarding the order. I didn't know swap wasn't `nothrow` as of C++11.