@LaurentGomila 2016-11-12
> The memcpy can be done as part of the initialization thus saving the internal memset

As mentioned in the previous comments, there was a reason for std::memset: it was much faster than vector::assign (which is most likely called by the equivalent constructor).