@binary1248 2016-11-12
@Fytch I knew there was a reason why the idiomatic way to shrink a vector pre-C++11 always swapped from the temporary... :roll_eyes:

@LaurentGomila I think you are mixing things up... `memset` is still called in the constructor now, as part of the filling with `T()` since only a size is passed. If T is a POD, the compiler is smart enough to do it all in one go with `memset` instead of looping over every element. If you initialize the vector directly from an array of PODs, the compiler is also smart enough to only write to the memory once, via the `memcpy`. A vector assign is slower whether or not the vector was initialized or not because it loops over every element, so that is not really what I was addressing. See [what the compiler generates at -O3](https://godbolt.org/g/7FB5ZF).