@Fytch 2016-11-12
As I am new to GitHub and collaborating on open source projects in general: How do I go about changing the code of this pull request without having to open a separate, second pull request? Should I just push another commit to my branch?