@victorlevasseur 2016-11-12
Yes

Le sam. 12 nov. 2016 12:52, Fytch <notifications@github.com> a écrit :

> As I am new to GitHub and collaborating on open source projects in
> general: How do I go about changing the code of this pull request without
> having to open a separate, second pull request? Should I just push another
> commit to my branch?
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/SFML/SFML/pull/1166#issuecomment-260117781>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AA63oLyd0fmaESP184lgFhgPVbZtvT-Jks5q9ahzgaJpZM4Kvo4M>
> .
>