@LaurentGomila 2016-11-12
> A vector assign is slower whether or not the vector was initialized or not because it loops over every element

Why in this specific case the compiler would not be as smart as in the constructor, and not use std::memcpy as well?