@binary1248 2016-11-12
[Maybe because -O3 was not used when initially benchmarking assign?](https://godbolt.org/g/zQFYvT) 😁