@LaurentGomila 2016-11-12
It was Visual Studio (probably an old version at that time), with the standard options generated by CMake in Release configuration.