@binary1248 2016-11-12
@Fytch Can you squash the commits into a single one?