@Fytch 2016-11-12
On an unrelated note: `std::vector<Uint8>` turns up quite often in Image.cpp. Why not make it a `typedef` inside of `sf::Image`'s class body in Image.hpp and use that `typedef` instead? It would live up to the DRY principle. The same goes for quite a lot of adapter classes.