@binary1248 2016-11-12
SFMLBuildMaster: Build this please.