@LaurentGomila 2016-11-13
I don't know what we would gain with an alias for `std::vector<Uint8>`. It's not long or hard to type, not likely to change in the future, and at least we know exactly what we deal with in the implementation.