@eXpl0it3r 2016-11-21
Probably not the best, but here some simple benchmarking code.

```cpp
#include <SFML/Graphics.hpp>
#include <iostream>

int main()
{
    const unsigned int ITERATIONS = 1000;
    const unsigned int WIDTH = 1920;
    const unsigned int HEIGHT = 1080;

    // Create a new sf::Image
    {
        sf::Clock clock;
        sf::Time average;

        for (unsigned int i = 0; i < ITERATIONS; ++i)
        {
            sf::Image image;

            clock.restart();
            image.create(WIDTH, HEIGHT, sf::Color::Green);
            average += clock.getElapsedTime();
        }

        std::cout << "New: " << (static_cast<float>(average.asMicroseconds()) / static_cast<float>(ITERATIONS)) << "\n";
    }

    // Re-use of sf::Image
    {
        sf::Clock clock;
        sf::Time average;

        sf::Image image;
        image.create(WIDTH, HEIGHT, sf::Color::Green);

        for (unsigned int i = 0; i < ITERATIONS; ++i)
        {
            clock.restart();
            image.create(WIDTH, HEIGHT, sf::Color::Green);
            average += clock.getElapsedTime();
        }

        std::cout << "Re-use: " << (static_cast<float>(average.asMicroseconds()) / static_cast<float>(ITERATIONS)) << "\n";
    }


    std::vector<sf::Uint8> vec;
    vec.reserve(WIDTH * HEIGHT * 4);
    for(std::size_t i = 0; i < vec.size(); i += 4)
    {
        vec.push_back(0x00);
        vec.push_back(0xFF);
        vec.push_back(0x00);
        vec.push_back(0xFF);
    }

    // Create a new sf::Image
    {
        sf::Clock clock;
        sf::Time average;

        for (unsigned int i = 0; i < ITERATIONS; ++i)
        {
            sf::Image image;

            clock.restart();
            image.create(WIDTH, HEIGHT, vec.data());
            average += clock.getElapsedTime();
        }

        std::cout << "Array New: " << (static_cast<float>(average.asMicroseconds()) / static_cast<float>(ITERATIONS)) << "\n";
    }

    // Re-use of sf::Image
    {
        sf::Clock clock;
        sf::Time average;

        sf::Image image;
        image.create(WIDTH, HEIGHT, vec.data());

        for (unsigned int i = 0; i < ITERATIONS; ++i)
        {
            clock.restart();
            image.create(WIDTH, HEIGHT, sf::Color::Green);
            average += clock.getElapsedTime();
        }

        std::cout << "Array Re-use: " << (static_cast<float>(average.asMicroseconds()) / static_cast<float>(ITERATIONS)) << "\n";
    }
}
```

I can't really see a big difference using -O3 on an i5-4460 @ 3.2 GHz.

**master**

Type          | Average Time
------------- | -------------
Color New     | 3488.05μs
Color Re-use  |  979.852μs
Array New     | 3660.14μs
Array Re-use  |  997.509μs

**this pr**

Type          | Average Time
------------- | -------------
Color New     | 3591.06μs
Color Re-use  |  955.014μs
Array New     | 3721.27μs
Array Re-use  | 1039.37μs

But since this isn't purely about performance, but also about exception safety and given that theory behind it makes more sense, it's seem ready to be merged. 😄 

(For me personally it showed once again that spending time on (micro) optimizations is not necessarily worth it.)